requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.8.0
mlflow>=2.10
joblib>=1.3
xgboost>=2.0
fastapi>=0.110
uvicorn[standard]>=0.29
dvc>=3.50
PyYAML>=6.0
